+++ backend/src/main/java/com/kop/morning/domain/article/article/controller/service/ArticleService.java
@@ -1,10 +1,10 @@
-package com.kop.morning.domain.article.service;
+package com.kop.morning.domain.article.article.controller.service;
 
-import com.kop.morning.domain.article.dto.requestDto.ArticleResponseDto;
-import com.kop.morning.domain.article.dto.requestDto.ArticleUpdateRequestDto;
-import com.kop.morning.domain.article.dto.responseDto.ArticleRequestDto;
-import com.kop.morning.domain.article.entity.Article;
-import com.kop.morning.domain.article.repository.ArticleRepository;
+import com.kop.morning.domain.article.article.dto.requestDto.ArticleResponseDto;
+import com.kop.morning.domain.article.article.dto.requestDto.ArticleUpdateRequestDto;
+import com.kop.morning.domain.article.article.dto.responseDto.ArticleRequestDto;
+import com.kop.morning.domain.article.article.entity.Article;
+import com.kop.morning.domain.article.article.repository.ArticleRepository;
 import com.kop.morning.domain.member.entity.Member;
 import com.kop.morning.domain.member.repository.MemberRepository;
 import com.kop.morning.global.Utils.SecurityUtil;
