+++ backend/src/main/java/com/kop/morning/domain/article/article/dto/responseDto/ArticleRequestDto.java
@@ -1,4 +1,4 @@
-package com.kop.morning.domain.article.dto.responseDto;
+package com.kop.morning.domain.article.article.dto.responseDto;
 
 import lombok.Data;
