+++ backend/src/main/java/com/kop/morning/domain/article/article/repository/ArticleRepository.java
@@ -1,6 +1,6 @@
-package com.kop.morning.domain.article.repository;
+package com.kop.morning.domain.article.article.repository;
 
-import com.kop.morning.domain.article.entity.Article;
+import com.kop.morning.domain.article.article.entity.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArticleRepository extends JpaRepository<Article, Long> {
